move driver metriclog to test
